DiscordBot/requirements.txt
discord.py
rcon